requirements.txt
matplotlib
pandas
seaborn
scipy
numpy==1.20.0
tensorflow-gpu>=2.0
scikit-learn
scikit-image
biopython>=1.79
bioframe==0.2.0
cooler==0.8.11
cooltools==0.4.1
cython